Benjam - Change float into double, begin debug, but not finished

+++ First_Step/src/jus/stage/samples/DichotomicSearch.java
@@ -61,7 +61,7 @@ public class DichotomicSearch {
 
 	}
 
-	public static HashMap<Long, HashMap<float[], Long>> getObservation() {
+	public static HashMap<Long, HashMap<double[], Long>> getObservation() {
 		HashMap<Integer, Long> samples;
 		try {
 			samples = DichotomicSearch.getSamples(Parameters.nbSample);
